+++ src/api/routes/index.ts
@@ -3,15 +3,15 @@ import express from 'express';
 import { env } from '../../env.js';
 import { expressLogger } from '../../utilities/expressLogger.js';
 import { authConfig } from '../configs/auth.js';
-import { files } from '../controllers/files.js';
-import { me } from '../controllers/me.js';
-import { project } from '../controllers/projects.js';
-import { roles } from '../controllers/roles.js';
-import { users } from '../controllers/users.js';
+import { file } from './file.js';
+import { me } from './me.js';
+import { project } from './project.js';
+import { role } from './role.js';
+import { user } from './user.js';
 import { currentSession } from '../middleware/auth.js';
 import { corsMiddleware } from '../middleware/cors.js';
 
-const router = express.Router();
+export const router = express.Router();
 
 if (Boolean(env.CORS_ENABLED) === true) {
   router.use(corsMiddleware);
@@ -25,10 +25,8 @@ router.use('/auth/*', ExpressAuth(authConfig));
 
 router.get('/health', (_req, res) => res.send('ok'));
 
-router.use(users);
-router.use(roles);
+router.use(user);
+router.use(role);
 router.use(project);
 router.use(me);
-router.use(files);
-
-export const apiRouter = router;
+router.use(file);
