+++ src/express/api.ts
@@ -1,9 +1,9 @@
 import boxen from 'boxen';
 import chalk from 'chalk';
 import { Express } from 'express';
-import { assets } from '../api/controllers/assets.js';
 import { errorHandler } from '../api/middleware/errorHandler.js';
-import { apiRouter } from '../api/router/apiRouter.js';
+import { asset } from '../api/routes/asset.js';
+import { router } from '../api/routes/index.js';
 import { env } from '../env.js';
 import { logger } from '../utilities/logger.js';
 import { Output } from '../utilities/output.js';
@@ -12,8 +12,8 @@ export const initApiServer = async (app: Express) => {
   const port = env.SERVER_PORT;
   const host = env.SERVER_HOST;
 
-  app.use('/', assets);
-  app.use('/api', apiRouter);
+  app.use('/', asset);
+  app.use('/api', router);
   app.use(errorHandler);
 
   app
